pyproject.toml
[tool.poetry]
name = "leIA"
version = "1.0.0"
description = "Aplicação que usa modelos de inteligência artificial da openAI para transcrição de áudio e vídeo."
authors = ["André Rocha <andrecorumba@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^1.5.3"
openai-whisper = "^20230124"
streamlit = "^1.19.0"
streamlit-option-menu = "^0.3.2"
pydub = "^0.25.1"
fastapi = "^0.110.0"
uvicorn = "^0.28.1"

[tool.poetry.scripts]
my-script = "leia:main"


[tool.poetry.group.dev.dependencies]
ignr = "^2.2"
mkdocs = "^1.4.2"
mkdocs-material = "^9.1.3"
mkdocstrings = "^0.20.0"
mkdocstrings-python = "^0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
